improve calculation accuracy
Allow using float point arguments to avoid truncation.

Test: Add (4.9, 4.9) should return 10.

+++ calculator.h
@@ -4,8 +4,8 @@
 class Calculator
 {
     public:
-        int Add (int, int);
-        int Sub (int, int);
+        int Add (double, double);
+        int Sub (double, double);
 };
 
 #endif//CALCULATOR_H
